Separate GFX9 memory chart test suites
