+++ projects/rocprofiler-compute/tests/test_categories.yaml
@@ -35,6 +35,8 @@ test_categories:
       - test_torch_cpp_loader
       - test_import_guard
       - test_mem_chart
+      - test_mem_chart_gfx9
+      - test_gfx9_memory_chart_integration
       - test_native_tool_finder
       - test_metrics_common
       - test_tty
@@ -99,6 +101,8 @@ test_categories:
       - test_pc_sampling_profile
       - test_rocprofv3_avail_interface
       - test_mem_chart
+      - test_mem_chart_gfx9
+      - test_gfx9_memory_chart_integration
       - test_native_tool_finder
       - test_metrics_common
       - test_tty
@@ -167,6 +171,8 @@ test_categories:
       - test_pc_sampling_profile
       - test_rocprofv3_avail_interface
       - test_mem_chart
+      - test_mem_chart_gfx9
+      - test_gfx9_memory_chart_integration
       - test_native_tool_finder
       - test_metrics_common
       - test_tty
@@ -235,6 +241,8 @@ test_categories:
       - test_pc_sampling_profile
       - test_rocprofv3_avail_interface
       - test_mem_chart
+      - test_mem_chart_gfx9
+      - test_gfx9_memory_chart_integration
       - test_native_tool_finder
       - test_metrics_common
       - test_tty
